move some nose tests around
