+++ extension/test/server/shaky/test_collapse.py
